+++ analyzer/src/main/java/nl/basjes/parse/useragent/calculate/FieldCalculator.java
@@ -23,4 +23,9 @@ import java.io.Serializable;
 
 public interface FieldCalculator extends Serializable {
     void calculate(MutableUserAgent userAgent);
+
+    default String[] getDependencies() {
+        return new String[0];
+    }
+
 }